Move files to game-headed
